requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
mysql-connector-python==8.0.22
protobuf==3.13.0
requests==2.24.0
six==1.15.0
urllib3==1.25.11
